Fixed rendition change event.
