+++ MuxExoPlayer/build.gradle
@@ -52,7 +52,7 @@ android {
             java.srcDirs = ['./src/r2_11_1/java', './src/r2_9_6_and_later/java']
         }
         r2_11_1_ads {
-            java.srcDirs = ['./src/r2_11_1/java', './src/r2_11_1_and_later/java']
+            java.srcDirs = ['./src/r2_11_1/java', './src/r2_9_6_and_later/java']
         }
     }
