+++ MuxExoPlayer/src/r2_9_6_and_later/java/com/mux/stats/sdk/muxstats/MuxBaseExoPlayer.java
@@ -31,12 +31,17 @@ import com.mux.stats.sdk.core.MuxSDKViewOrientation;
 import com.mux.stats.sdk.core.events.EventBus;
 import com.mux.stats.sdk.core.events.IEvent;
 import com.mux.stats.sdk.core.events.InternalErrorEvent;
+import com.mux.stats.sdk.core.events.playback.AdPlayEvent;
+import com.mux.stats.sdk.core.events.playback.AdPlayingEvent;
+import com.mux.stats.sdk.core.events.playback.AdResponseEvent;
 import com.mux.stats.sdk.core.events.playback.EndedEvent;
 import com.mux.stats.sdk.core.events.playback.PauseEvent;
 import com.mux.stats.sdk.core.events.playback.PlayEvent;
 import com.mux.stats.sdk.core.events.playback.PlayingEvent;
 import com.mux.stats.sdk.core.events.playback.RenditionChangeEvent;
 import com.mux.stats.sdk.core.events.playback.RequestBandwidthEvent;
+import com.mux.stats.sdk.core.events.playback.SeekedEvent;
+import com.mux.stats.sdk.core.events.playback.SeekingEvent;
 import com.mux.stats.sdk.core.events.playback.TimeUpdateEvent;
 import com.mux.stats.sdk.core.model.BandwidthMetricData;
 import com.mux.stats.sdk.core.model.CustomerPlayerData;
@@ -52,7 +57,7 @@ import java.util.concurrent.atomic.AtomicLong;
 import static android.os.SystemClock.elapsedRealtime;
 
 public class MuxBaseExoPlayer extends EventBus implements IPlayerListener {
-    protected static final String TAG = "MuxStatsListener";
+    protected static final String TAG = "MuxBaseExoPlayer";
     // Error codes start at -1 as ExoPlaybackException codes start at 0 and go up.
     protected static final int ERROR_UNKNOWN = -1;
     protected static final int ERROR_DRM = -2;
@@ -73,10 +78,16 @@ public class MuxBaseExoPlayer extends EventBus implements IPlayerListener {
     protected int streamType = -1;
 
     public enum PlayerState {
-        BUFFERING, ERROR, PAUSED, PLAY, PLAYING, INIT, ENDED
+        BUFFERING, ERROR, PAUSED, PLAY, PLAYING, INIT, SEEKING, SEEKED, ENDED
     }
     protected PlayerState state;
     protected MuxStats muxStats;
+    protected AdsImaSDKListener imaListener;
+
+    protected boolean missedAfterAdsPlayEvent = false;
+    protected boolean missedAfterAdsPlayingEvent = false;
+    protected boolean missedAfterSeekingPlayingEvent = false;
+    protected Format selectedTrackFormat;
 
 
     MuxBaseExoPlayer(Context ctx, ExoPlayer player, String playerName, CustomerPlayerData customerPlayerData, CustomerVideoData customerVideoData, boolean sentryEnabled) {
@@ -88,6 +99,7 @@ public class MuxBaseExoPlayer extends EventBus implements IPlayerListener {
         muxStats = new MuxStats(this, playerName, customerPlayerData, customerVideoData, sentryEnabled);
         addListener(muxStats);
         Player.VideoComponent lDecCount = player.getVideoComponent();
+
         playerHandler = new ExoPlayerHandler(player.getApplicationLooper(), player);
         lDecCount.setVideoFrameMetadataListener(new VideoFrameMetadataListener() {
             // As of r2.11.x, the signature for this callback has changed. These are not annotated as @Overrides in
@@ -100,30 +112,6 @@ public class MuxBaseExoPlayer extends EventBus implements IPlayerListener {
                 playerHandler.obtainMessage(ExoPlayerHandler.UPDATE_PLAYER_CURRENT_POSITION).sendToTarget();
             }
         });
-    }
-
-    /**
-     * Get the instance of the IMA SDK Listener for tracking ads running through Google's
-     * IMA SDK within your application.
-     *
-     * @deprecated
-     * This method is no longer the preferred method to track Ad performance with
-     * Google's IMA SDK.
-     * <p> Use {@link MuxBaseExoPlayer#monitorImaAdsLoader(AdsLoader)} instead.
-     * @return the IMA SDK Listener
-     * @throws
-     */
-    @Deprecated
-    public AdsImaSDKListener getIMASdkListener() {
-        try {
-            // Let's just check one of them
-            Class.forName("com.google.ads.interactivemedia.v3.api.Ad");
-            Class.forName("com.google.ads.interactivemedia.v3.api.AdErrorEvent");
-            Class.forName("com.google.ads.interactivemedia.v3.api.AdEvent");
-            return new AdsImaSDKListener(this);
-        } catch (ClassNotFoundException cnfe) {
-            throw new IllegalStateException("IMA SDK Modules not found");
-        }
     }
 
     /**
@@ -151,7 +139,7 @@ public class MuxBaseExoPlayer extends EventBus implements IPlayerListener {
 
                     // Set up the ad events that we want to use
                     AdsManager adsManager = adsManagerLoadedEvent.getAdsManager();
-                    AdsImaSDKListener imaListener = new AdsImaSDKListener(baseExoPlayer);
+                    imaListener = new AdsImaSDKListener(baseExoPlayer);
                     // Attach mux event and error event listeners.
                     adsManager.addAdErrorListener(imaListener);
                     adsManager.addAdEventListener(imaListener);
@@ -164,7 +152,22 @@ public class MuxBaseExoPlayer extends EventBus implements IPlayerListener {
         }
     }
 
-    @SuppressWarnings("unused")
+    public void setAdsListener(AdsImaSDKListener listener) {
+        if (listener != null) {
+            imaListener = listener;
+            imaListener.setExoPlayerListener(this);
+        }
+    }
+
+    public boolean isMissedAfterAdsPlayEvent() {
+        return missedAfterAdsPlayEvent;
+    }
+
+    public boolean isMissedAfterAdsPlayingEvent() {
+        return missedAfterAdsPlayingEvent;
+    }
+	
+	@SuppressWarnings("unused")
     public void updateCustomerData(CustomerPlayerData customPlayerData, CustomerVideoData customVideoData) {
         muxStats.updateCustomerData(customPlayerData, customVideoData);
     }
@@ -320,20 +323,45 @@ public class MuxBaseExoPlayer extends EventBus implements IPlayerListener {
     }
 
     protected void play() {
+        if (state == PlayerState.PLAY) {
+            // Ignore, redundant
+            return;
+        }
         state = PlayerState.PLAY;
         dispatch(new PlayEvent(null));
+        missedAfterAdsPlayEvent = false;
     }
 
     protected void playing() {
-        if (state == PlayerState.PAUSED) {
-            play();
+        if (state == PlayerState.PLAYING) {
+            // Ignore, redundant
+            return;
         }
         state = PlayerState.PLAYING;
         dispatch(new PlayingEvent(null));
+        missedAfterAdsPlayingEvent = false;
+    }
+
+    protected void seekStarted() {
+        state = PlayerState.SEEKING;
+        dispatch(new SeekingEvent(null));
+    }
+
+    protected void seekEnded() {
+        synchronized (this) {
+            state = PlayerState.SEEKED;
+            dispatch(new SeekedEvent(null));
+            if (missedAfterSeekingPlayingEvent) {
+                playing();
+            }
+            missedAfterSeekingPlayingEvent = false;
+        }
     }
 
     protected void ended() {
-        dispatch(new PauseEvent(null));
+        if (state != PlayerState.PAUSED) {
+            dispatch(new PauseEvent(null));
+        }
         dispatch(new EndedEvent(null));
         state = PlayerState.ENDED;
     }
@@ -349,6 +377,10 @@ public class MuxBaseExoPlayer extends EventBus implements IPlayerListener {
 
     protected void handleRenditionChange(Format format) {
         if (format != null) {
+            if (format == selectedTrackFormat) {
+                // same track, ignore
+                return;
+            }
             sourceAdvertisedBitrate = format.bitrate;
             if (format.frameRate > 0) {
                 sourceAdvertisedFramerate = format.frameRate;
@@ -357,6 +389,7 @@ public class MuxBaseExoPlayer extends EventBus implements IPlayerListener {
             sourceHeight = format.height;
             RenditionChangeEvent event = new RenditionChangeEvent(null);
             dispatch(event);
+            selectedTrackFormat = format;
         }
     }
